requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
distlib==0.3.4
Django==4.0.3
django-ckeditor==6.2.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-etc==1.3.1
django-hitcount==1.3.5
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-rosetta==0.9.8
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.1.0
drf-yasg==1.21.3
filelock==3.6.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.0
platformdirs==2.5.2
polib==1.1.1
pycparser==2.21
PyJWT==2.3.0
pyparsing==3.0.9
python-gettext==4.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
stripe==3.0.0
twilio==7.8.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenvwrapper-win==1.2.7
